requirements.txt
# FastAPI and ASGI server
fastapi==0.103.2
uvicorn[standard]==0.23.2

# Database
sqlalchemy==2.0.20
alembic==1.12.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment
python-dotenv==1.0.0
pydantic-settings==2.0.3

# Validation
pydantic==2.4.2
email-validator==2.0.0

# HTTP client
httpx==0.24.1

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.7.0
isort==5.12.0
flake8==6.0.0

# Telegram Bot (for future integration)
python-telegram-bot==20.4

# Additional utilities
python-dateutil==2.8.2

# Communication system dependencies
celery==5.3.2
redis==5.0.0
jinja2==3.1.2
websockets==11.0.3